// src/app/models/admin/capphephoatdongkhoangsan/cptanthukhoangsan/cptanthukhoangsan.model.ts
export class OutputCpTanThuKhoangSanModel {
  idcappheptanthu: string;
  idkhuvuctanthu: string;
  tenkhuvuctanthu: string;
  diadiem: string;
  dientichkhaithac: number;
  truluongdiachat: number;
  truluongkhaithac: number;
  thoihankhaithac: number;
  ngaybdkhaithac: string;
  ngayktkhaithac: string;
  phuongphapkhaithac: number;
  congsuatkhaithac: number;
  mucsaukhaithactu: number;
  mucsaukhaithacden: number;
  dangkhoangsan: number;
  donvitruluong: string;
  donvicongsuat: string;
  donvidientich: string;
  donvithoihan: string;
  donvichieusau: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  idgiayphep: string;
  siteid: string;
}

export class InputCpTanThuKhoangSanModel {
  idcappheptanthu: string;
  tenkhuvuctanthu: string;
  diadiem: string;
  dientichkhaithac: number;
  truluongdiachat: number;
  truluongkhaithac: number;
  thoihankhaithac: number;
  ngaybdkhaithac: string;
  ngayktkhaithac: string;
  phuongphapkhaithac: number;
  congsuatkhaithac: number;
  mucsaukhaithactu: number;
  mucsaukhaithacden: number;
  dangkhoangsan: number;
  donvitruluong: string;
  donvicongsuat: string;
  donvidientich: string;
  donvithoihan: string;
  donvichieusau: string;
  hequychieu: string;
  idgiayphep: string;
  siteid: string;
}